Fix MHeapAlignDown() when MHEAP_BLOCK_ALIGN is defined to integer constant
On platforms where SIZE_MAX > UINT_MAX, if MHEAP_BLOCK_ALIGN was defined
 to something with type int or unsigned int, MHeapAlignDown() would strip
 upper bits of its argument.

+++ mheap.h
@@ -67,7 +67,7 @@ typedef struct {
 
 
 // Helper macros to align block sizes
-#define MHeapAlignDown(size) ((size) & ~(MHEAP_BLOCK_ALIGN - 1u))
+#define MHeapAlignDown(size) ((size) & ~((size_t)MHEAP_BLOCK_ALIGN - 1u))
 #define MHeapAlignUp(size) MHeapAlignDown((size) + (MHEAP_BLOCK_ALIGN - 1u))
